The `refactor-and-bugfix` branch was broken and contained many commits that were not published to the remote. Therefore, it had to be recovered, resulting in the loss of all previous commit history.
