+++ examples/hierarchical_example.py
@@ -1,13 +1,13 @@
 import os
 
-from streamlit_langgraph import UIConfig, LangGraphChat, load_agents_from_yaml
+from streamlit_langgraph import AgentManager, UIConfig, LangGraphChat
 from streamlit_langgraph.workflow import WorkflowBuilder, SupervisorTeam
 
 def create_hierarchical_workflow_example():
     """Create a hierarchical workflow with multiple supervisor teams."""
     
     config_path = os.path.join(os.path.dirname(__file__), "./configs/hierarchical.yaml")
-    agents = load_agents_from_yaml(config_path)
+    agents = AgentManager.load_from_yaml(config_path)
 
     # Unpack agents
     project_manager = agents[0]
